requirements.txt
django==5.0.2
django-debug-toolbar==4.3.0
celery==5.3.6
flower==2.0.1
django-redis==5.4.0
tornado==6.4
django-crontab==0.7.1
django-cron==0.6.0
gunicorn==21.2.0
rich==13.7.1
python-decouple==3.8
django-cors-headers==4.3.1
pillow==10.2.0
django-taggit==5.0.1
flake8 == 7.0.0 
